Trello Updates & Assist
I am monitoring the Trello boards to see if everyone is up to date if anyone requires help, the tickets have written descriptions and comments, and that all tickets are done in a timely manner. 

I have an image here as well which shows my updates, where I have added comments and descriptions to leave records of what I have done.